@@ -19,6 +19,7 @@ use League\Uri\Exceptions\SyntaxError;
 use League\Uri\Exceptions\TemplateCanNotBeExpanded;
 use League\Uri\UriTemplate\Template;
 use League\Uri\UriTemplate\VariableBag;
+use TypeError;
 
 /**
  * Defines the URI Template syntax and the process for expanding a URI Template into a URI reference.
@@ -33,20 +34,13 @@ use League\Uri\UriTemplate\VariableBag;
  */
 final class UriTemplate
 {
-    /**
-     * @var Template
-     */
-    private $template;
-
-    /**
-     * @var VariableBag
-     */
-    private $defaultVariables;
+    private Template $template;
+    private VariableBag $defaultVariables;
 
     /**
      * @param object|string $template a string or an object with the __toString method
      *
-     * @throws \TypeError               if the template is not a string or an object with the __toString method
+     * @throws TypeError                if the template is not a string or an object with the __toString method
      * @throws SyntaxError              if the template syntax is invalid
      * @throws TemplateCanNotBeExpanded if the template variables are invalid
      */
@@ -119,10 +113,10 @@ final class UriTemplate
      */
     public function withDefaultVariables(array $defaultDefaultVariables): self
     {
-        $clone = clone $this;
-        $clone->defaultVariables = $this->filterVariables($defaultDefaultVariables);
-
-        return $clone;
+        return new self(
+            $this->template->toString(),
+            $this->filterVariables($defaultDefaultVariables)->all()
+        );
     }
 
     /**
@@ -131,10 +125,10 @@ final class UriTemplate
      */
     public function expand(array $variables = []): UriInterface
     {
-        $uriString = $this->template->expand(
-            $this->filterVariables($variables)->replace($this->defaultVariables)
+        return Uri::createFromString(
+            $this->template->expand(
+                $this->filterVariables($variables)->replace($this->defaultVariables)
+            )
         );
-
-        return Uri::createFromString($uriString);
     }
 }
