@@ -64,30 +64,13 @@ final class Expression
         '&' => ['prefix' => '&', 'joiner' => '&', 'query' => true],
     ];
 
-    /**
-     * @var string
-     */
-    private $operator;
-
-    /**
-     * @var string
-     */
-    private $joiner;
-
-    /**
-     * @var array<VarSpecifier>
-     */
-    private $varSpecifiers;
-
-    /**
-     * @var array<string>
-     */
-    private $variableNames;
-
-    /**
-     * @var string
-     */
-    private $expressionString;
+    private string $operator;
+    /** @var array<VarSpecifier> */
+    private array $varSpecifiers;
+    private string $joiner;
+    /** @var array<string> */
+    private array $variableNames;
+    private string $expressionString;
 
     private function __construct(string $operator, VarSpecifier ...$varSpecifiers)
     {
@@ -103,20 +86,18 @@ final class Expression
      */
     private function setVariableNames(): array
     {
-        $mapper = static function (VarSpecifier $varSpecifier): string {
-            return $varSpecifier->name();
-        };
-
-        return array_unique(array_map($mapper, $this->varSpecifiers));
+        return array_unique(array_map(
+            static fn (VarSpecifier $varSpecifier): string => $varSpecifier->name(),
+            $this->varSpecifiers
+        ));
     }
 
     private function setExpressionString(): string
     {
-        $mapper = static function (VarSpecifier $variable): string {
-            return $variable->toString();
-        };
-
-        $varSpecifierString = implode(',', array_map($mapper, $this->varSpecifiers));
+        $varSpecifierString = implode(',', array_map(
+            static fn (VarSpecifier $variable): string => $variable->toString(),
+            $this->varSpecifiers
+        ));
 
         return '{'.$this->operator.$varSpecifierString.'}';
     }
@@ -146,11 +127,10 @@ final class Expression
             throw new SyntaxError('The operator used in the expression "'.$expression.'" is reserved.');
         }
 
-        $mapper = static function (string $varSpec): VarSpecifier {
-            return VarSpecifier::createFromString($varSpec);
-        };
-
-        return new Expression($parts['operator'], ...array_map($mapper, explode(',', $parts['variables'])));
+        return new Expression($parts['operator'], ...array_map(
+            static fn (string $varSpec): VarSpecifier => VarSpecifier::createFromString($varSpec),
+            explode(',', $parts['variables'])
+        ));
     }
 
     /**
@@ -177,11 +157,7 @@ final class Expression
             $parts[] = $this->replace($varSpecifier, $variables);
         }
 
-        $nullFilter = static function ($value): bool {
-            return '' !== $value;
-        };
-
-        $expanded = implode($this->joiner, array_filter($parts, $nullFilter));
+        $expanded = implode($this->joiner, array_filter($parts, static fn ($value): bool => '' !== $value));
         if ('' === $expanded) {
             return $expanded;
         }
