@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace League\Uri\UriTemplate;
 
 use League\Uri\Exceptions\TemplateCanNotBeExpanded;
+use TypeError;
 use function gettype;
 use function is_array;
 use function is_bool;
@@ -27,10 +28,10 @@ final class VariableBag
     /**
      * @var array<string,string|array<string>>
      */
-    private $variables = [];
+    private array $variables = [];
 
     /**
-     * @param iterable<string,mixed> $variables
+     * @param iterable<string,string|bool|int|float|array<string|bool|int|float>> $variables
      */
     public function __construct(iterable $variables = [])
     {
@@ -63,7 +64,7 @@ final class VariableBag
     }
 
     /**
-     * @param string|array<string> $value
+     * @param string|bool|int|float|array<string|bool|int|float> $value
      */
     public function assign(string $name, $value): void
     {
@@ -88,7 +89,7 @@ final class VariableBag
         }
 
         if (!is_array($value)) {
-            throw new \TypeError(sprintf('The variable '.$name.' must be NULL, a scalar or a stringable object `%s` given', gettype($value)));
+            throw new TypeError(sprintf('The variable '.$name.' must be NULL, a scalar or a stringable object `%s` given', gettype($value)));
         }
 
         if (!$isNestedListAllowed) {
@@ -108,9 +109,6 @@ final class VariableBag
      */
     public function replace(VariableBag $variables): self
     {
-        $instance = clone $this;
-        $instance->variables += $variables->variables;
-
-        return $instance;
+        return new self($this->variables + $variables->variables);
     }
 }
