@@ -15,6 +15,7 @@ namespace League\Uri;
 
 use League\Uri\Contracts\UriInterface;
 use Psr\Http\Message\UriInterface as Psr7UriInterface;
+use TypeError;
 use function array_pop;
 use function array_reduce;
 use function count;
@@ -94,12 +95,12 @@ final class UriResolver
      *
      * @param mixed $uri an URI object
      *
-     * @throws \TypeError if the URI object does not implements the supported interfaces.
+     * @throws TypeError if the URI object does not implements the supported interfaces.
      */
     private static function filterUri($uri): void
     {
         if (!$uri instanceof UriInterface && !$uri instanceof Psr7UriInterface) {
-            throw new \TypeError(sprintf('The uri must be a valid URI object received `%s`', gettype($uri)));
+            throw new TypeError(sprintf('The uri must be a valid URI object received `%s`', gettype($uri)));
         }
     }
 
@@ -270,9 +271,9 @@ final class UriResolver
     /**
      * Filter the URI object.
      *
-     * @param null|mixed $uri
+     * @param Psr7UriInterface|UriInterface $uri
      *
-     * @throws \TypeError if the URI object does not implements the supported interfaces.
+     * @throws TypeError if the URI object does not implements the supported interfaces.
      *
      * @return Psr7UriInterface|UriInterface
      */
